加入JM与U图

+++ web/src/main/java/com/jeesite/modules/jm/reliability/base/compareModel.java
@@ -19,7 +19,7 @@ public abstract class compareModel {
 		return this.name+" "+mo.getName();
 	}
 	
-	public double getResultNum() {
+	public int getResultNum() {
         return resultNum;
     }
